+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added new optional configuration option `origin` that is used as a value for the header `X-Amp-Origin`.
 
+### Changed
+
+- Changed the default templates - images are now rendered inside `<picture>` with `<source>` tags inside if the API returns data for them.
+
 ## [1.2.0] - 2023-08-14
 
 ### Added